requirements.txt
# RAG Pipeline Requirements
anthropic==0.39.0
chromadb==0.5.23
pypdf==5.1.0
sentence-transformers==3.3.1
streamlit==1.40.1
python-dotenv==1.0.1
numpy==1.26.4
pandas==2.2.3
